asim: move exchange into gossip pkg
This patch moves `exchange.go` and `exchange_test.go` into a separate
pkg. These files were previously tracked under `asim/state`. This
separation enforces stricter interaction between the state and gossip
components within asim for the purpose of modularity.

Release note: None
