+++ pkg/BUILD.bazel
@@ -185,6 +185,7 @@ ALL_TESTS = [
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl_test",
     "//pkg/kv/kvserver/allocator/storepool:storepool_test",
     "//pkg/kv/kvserver/apply:apply_test",
+    "//pkg/kv/kvserver/asim/gossip:gossip_test",
     "//pkg/kv/kvserver/asim/op:op_test",
     "//pkg/kv/kvserver/asim/queue:queue_test",
     "//pkg/kv/kvserver/asim/state:state_test",
@@ -1155,6 +1156,8 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/apply:apply",
     "//pkg/kv/kvserver/apply:apply_test",
     "//pkg/kv/kvserver/asim/config:config",
+    "//pkg/kv/kvserver/asim/gossip:gossip",
+    "//pkg/kv/kvserver/asim/gossip:gossip_test",
     "//pkg/kv/kvserver/asim/op:op",
     "//pkg/kv/kvserver/asim/op:op_test",
     "//pkg/kv/kvserver/asim/queue:queue",
@@ -2541,6 +2544,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/kv/kvserver/apply:get_x_data",
     "//pkg/kv/kvserver/asim:get_x_data",
     "//pkg/kv/kvserver/asim/config:get_x_data",
+    "//pkg/kv/kvserver/asim/gossip:get_x_data",
     "//pkg/kv/kvserver/asim/op:get_x_data",
     "//pkg/kv/kvserver/asim/queue:get_x_data",
     "//pkg/kv/kvserver/asim/state:get_x_data",
