+++ pkg/kv/kvserver/asim/gossip/BUILD.bazel
@@ -1,0 +1,27 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "gossip",
+    srcs = ["exchange.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/gossip",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/kv/kvserver/allocator/storepool",
+        "//pkg/roachpb",
+    ],
+)
+
+go_test(
+    name = "gossip_test",
+    srcs = ["exchange_test.go"],
+    args = ["-test.timeout=295s"],
+    embed = [":gossip"],
+    deps = [
+        "//pkg/kv/kvserver/asim/state",
+        "//pkg/roachpb",
+        "@com_github_stretchr_testify//require",
+    ],
+)
+
+get_x_data(name = "get_x_data")
