+++ pkg/kv/kvserver/asim/gossip/exchange.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package state
+package gossip
 
 import (
 	"fmt"
