+++ pkg/kv/kvserver/asim/queue/BUILD.bazel
@@ -29,6 +29,7 @@ go_test(
     embed = [":queue"],
     deps = [
         "//pkg/kv/kvserver/asim/config",
+        "//pkg/kv/kvserver/asim/gossip",
         "//pkg/kv/kvserver/asim/state",
         "//pkg/kv/kvserver/asim/workload",
         "//pkg/roachpb",
