+++ pkg/kv/kvserver/asim/storerebalancer/BUILD.bazel
@@ -38,6 +38,7 @@ go_test(
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/allocator/load",
         "//pkg/kv/kvserver/asim/config",
+        "//pkg/kv/kvserver/asim/gossip",
         "//pkg/kv/kvserver/asim/op",
         "//pkg/kv/kvserver/asim/state",
         "//pkg/roachpb",
